% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{construir_url_bball_ref}
\alias{construir_url_bball_ref}
\title{Construir URLs de Basketball Reference}
\usage{
construir_url_bball_ref(tipo, parametros)
}
\arguments{
\item{tipo}{Tipo de URL: "equipo_temporada", "jugador", "temporada_stats", "playoffs"}

\item{parametros}{Lista con parámetros necesarios según el tipo}
}
\value{
String con la URL completa
}
\description{
Función genérica para construir URLs de diferentes secciones de Basketball Reference
}
\examples{
\dontrun{
url <- construir_url_bball_ref("equipo_temporada", list(codigo = "LAL", year = 2020))
}

}
\keyword{internal}
